jcr: moved code for timer signals to the dedicated file
